Refactor to layered architecture
