+++ Adapters/Repositories/HttpAirportRepository.cs
@@ -1,40 +1,35 @@
 using System.Net;
 using System.Net.Http.Json;
-using DistanceService.Application.Interfaces;
+using DistanceService.Adapters.Options;
+using DistanceService.Adapters.Ports;
 using DistanceService.Domain.Entities;
-using DistanceService.Infrastructure.Options;
 using Microsoft.Extensions.Options;
 
-namespace DistanceService.Infrastructure.Repositories;
+namespace DistanceService.Adapters.Repositories;
 
 public sealed class HttpAirportRepository : IAirportRepository
 {
     private readonly HttpClient _httpClient;
     private readonly ICacheService<Airport> _cache;
-    private readonly AirportApiOptions _options;
+    private readonly AdaptersOptions _options;
 
     public HttpAirportRepository(HttpClient httpClient,
                                  ICacheService<Airport> cache,
-                                 IOptions<AirportApiOptions> options)
+                                 IOptions<AdaptersOptions> options)
     {
         _httpClient = httpClient ?? throw new ArgumentNullException(nameof(httpClient));
         _cache = cache ?? throw new ArgumentNullException(nameof(cache));
-        if (options == null) throw new ArgumentNullException(nameof(options));
-        _options = options.Value;
+        _options = options?.Value ?? throw new ArgumentNullException(nameof(options));
     }
 
     public async Task<Airport?> GetAirportAsync(string iata, CancellationToken cancellationToken = default)
     {
         if (string.IsNullOrWhiteSpace(iata))
-        {
             throw new ArgumentException("IATA code is required", nameof(iata));
-        }
         iata = iata.Trim().ToUpperInvariant();
 
         if (_cache.TryGet(iata, out var cached))
-        {
             return cached;
-        }
 
         var requestUri = _options.BaseUrl + Uri.EscapeDataString(iata);
 
@@ -43,9 +38,7 @@ public sealed class HttpAirportRepository : IAirportRepository
             .ConfigureAwait(false);
 
         if (response.StatusCode == HttpStatusCode.NotFound)
-        {
             return null;
-        }
 
         response.EnsureSuccessStatusCode();
 
@@ -54,9 +47,7 @@ public sealed class HttpAirportRepository : IAirportRepository
             .ConfigureAwait(false);
 
         if (airport?.Location is null)
-        {
             return null;
-        }
 
         airport.Iata = iata;
