+++ Adapters/Services/BoundedMemoryCacheService.cs
@@ -1,26 +1,24 @@
-using DistanceService.Application.Interfaces;
-using DistanceService.Infrastructure.Options;
+using DistanceService.Adapters.Options;
+using DistanceService.Adapters.Ports;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Options;
 
-namespace DistanceService.Infrastructure.Services;
+namespace DistanceService.Adapters.Services;
 
 public sealed class BoundedMemoryCacheService<T> : ICacheService<T>, IDisposable
 {
     private readonly MemoryCache _cache;
 
-    public BoundedMemoryCacheService(IOptions<CacheOptions> options)
+    public BoundedMemoryCacheService(IOptions<AdaptersOptions> options)
     {
         if (options == null) throw new ArgumentNullException(nameof(options));
-        var maxEntries = options.Value.MaxEntries;
+        var maxEntries = options.Value.CacheMaxEntries;
         if (maxEntries <= 0) maxEntries = 1000;
         _cache = new MemoryCache(new MemoryCacheOptions { SizeLimit = maxEntries });
     }
 
     public bool TryGet(string key, out T value)
-    {
-        return _cache.TryGetValue(key, out value!);
-    }
+        => _cache.TryGetValue(key, out value!);
 
     public void Set(string key, T value, TimeSpan expiration)
     {
@@ -32,8 +30,5 @@ public sealed class BoundedMemoryCacheService<T> : ICacheService<T>, IDisposable
         _cache.Set(key, value, entryOptions);
     }
 
-    public void Dispose()
-    {
-        _cache.Dispose();
-    }
+    public void Dispose() => _cache.Dispose();
 }
