--- Application/Options/DistanceOptions.cs
@@ -1,6 +1,0 @@
-namespace DistanceService.Application.Options;
-
-public class DistanceOptions
-{
-    public double EarthRadiusMiles { get; set; } = 3958.8;
-}
