+++ DistanceService.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 17.10.35122.118
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DistanceService", "DistanceService.csproj", "{171491A0-6BD2-4111-8E3E-EBFC8B07E3E1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{72B733C1-8434-4190-AF4F-FD0F756370DE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adapters", "Adapters\Adapters.csproj", "{D3ADC0EF-4871-4A76-8139-922714A3EB60}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +19,14 @@ Global
 		{171491A0-6BD2-4111-8E3E-EBFC8B07E3E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{171491A0-6BD2-4111-8E3E-EBFC8B07E3E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{171491A0-6BD2-4111-8E3E-EBFC8B07E3E1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{72B733C1-8434-4190-AF4F-FD0F756370DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{72B733C1-8434-4190-AF4F-FD0F756370DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{72B733C1-8434-4190-AF4F-FD0F756370DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{72B733C1-8434-4190-AF4F-FD0F756370DE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D3ADC0EF-4871-4A76-8139-922714A3EB60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D3ADC0EF-4871-4A76-8139-922714A3EB60}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D3ADC0EF-4871-4A76-8139-922714A3EB60}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D3ADC0EF-4871-4A76-8139-922714A3EB60}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
