--- Infrastructure/Options/CacheOptions.cs
@@ -1,6 +1,0 @@
-namespace DistanceService.Infrastructure.Options;
-
-public class CacheOptions
-{
-    public int MaxEntries { get; set; } = 1000;
-}
