+++ Program.cs
@@ -1,5 +1,4 @@
-using DistanceService.Application;
-using DistanceService.Infrastructure;
+using DistanceService.Adapters;
 using DistanceService.Presentation.Authentication;
 using DistanceService.Presentation.Middleware;
 using DistanceService.Presentation.Options;
@@ -10,8 +9,7 @@ var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddControllers();
 
-builder.Services.AddApplication(builder.Configuration);
-builder.Services.AddInfrastructure(builder.Configuration);
+builder.Services.AddAdapters(builder.Configuration);
 builder.Services.Configure<AuthOptions>(builder.Configuration.GetSection("Auth"));
 
 builder.Services.AddAuthentication("Bearer")
