+++ appsettings.json
@@ -1,13 +1,8 @@
 {
-  "AirportApi": {
+  "Adapters": {
     "BaseUrl": "https://places-dev.continent.ru/airports/",
-    "CacheDurationMinutes": 60
-  },
-  "Distance": {
-    "EarthRadiusMiles": 3958.8
-  },
-  "Cache": {
-    "MaxEntries": 1000
+    "CacheDurationMinutes": 60,
+    "CacheMaxEntries": 1000
   },
   "Auth": {
     "Token": "test"
